refactor: fix header and nav menu
